+++ CHANGELOG.md
@@ -1,2 +1,20 @@
+# Version 1.0.3
+
+Fixed issues where default values were being ignored instead of being detected as a change
+
+Added a list of the changes to the dialog box requesting a sync.
+
+Added a check box to ignore changes until the next setting change.
+
+Added a button in the settings to reset the ignore options.
+
+Fixed an issue with showing player changes to the GM
+
+Fixed an issue where closing the settings config while viewing a player would show the Games master as the one being viewed when the dialog opens again, but still show player data.
+
+# Version 1.0.2
+
+Fixed an issue with logging changes being sync'd
+
 # Version 1.0.0
 Beta version in development